Moves txn code into a new endpoint, not specific to KV.
